Formula for goodhosts 1.1.0
We recommend using goodhosts in the payments-service README to modify
/etc/hosts. We add some GC-specific hostnames to localhost for
convenience.

goodhosts' recommended installation process involves installing the go
binary for your CPU architecture and placing it in /usr/local/bin. We
could create a formula for goodhosts in our internal GC tap to make
setting up the payments-service a bit easier. As we're also building
the application from source, users don't need to worry about choosing
the correct Go binary from the releases page.

Installation would then be:

brew tap gocardless/homebrew-taps
brew install goodhosts

I tried to add goodhosts to the core homebrew tap, but the maintainers
closed the issue citing the project is too small, and advised to host
it in another tap:

https://github.com/Homebrew/homebrew-core/pull/95450