Remove unused ruby bits
We don't run rubocop, so this is unused.

The versions are also really old.

Most of the ruby code in this repo is generated by goreleaser, so applying
additional/different rubocop config will probably conflict and not add much
value.

Instead, we can rely on the tests added to CI to prevent major issues.

`brew style` could be an option as this checks against the homebrew style
guide, but again, most of our code is generated and so not easy/possible to
change (I tried).